backend/pyproject.toml
[project]
name = "web-presence-scanner"
version = "0.1.0"
description = "Web presence scanner API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis>=5.0",
    "arq>=0.26",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "sse-starlette>=2.2",
    "PyJWT[crypto]>=2.9",
]
